+++ lib/internal/bootstrap/pre_execution.js
@@ -25,7 +25,7 @@ function prepareMainThreadExecution(expandArgv1 = false) {
   // prepareStackTrace method, replacing the default in errors.js.
   if (getOptionValue('--enable-source-maps')) {
     const { prepareStackTrace } =
-      require('internal/source_map/source_map_cache');
+      require('internal/source_map/prepare_stack_trace');
     const { setPrepareStackTraceCallback } = internalBinding('errors');
     setPrepareStackTraceCallback(prepareStackTrace);
   }
