+++ lib/internal/source_map/source_map_cache.js
@@ -17,7 +17,6 @@ const cjsSourceMapCache = new WeakMap();
 // on filenames.
 const esmSourceMapCache = new Map();
 const { fileURLToPath, URL } = require('url');
-const { overrideStackTrace } = require('internal/errors');
 
 let experimentalSourceMaps;
 function maybeCacheSourceMap(filename, content, cjsModuleInstance) {
@@ -38,18 +37,22 @@ function maybeCacheSourceMap(filename, content, cjsModuleInstance) {
 
   const match = content.match(/\/[*/]#\s+sourceMappingURL=(?<sourceMappingURL>[^\s]+)/);
   if (match) {
+    const data = dataFromUrl(basePath, match.groups.sourceMappingURL);
+    const url = data ? null : match.groups.sourceMappingURL;
     if (cjsModuleInstance) {
       cjsSourceMapCache.set(cjsModuleInstance, {
         filename,
-        url: match.groups.sourceMappingURL,
-        data: dataFromUrl(basePath, match.groups.sourceMappingURL)
+        lineLengths: lineLengths(content),
+        data,
+        url
       });
     } else {
       // If there is no cjsModuleInstance assume we are in a
       // "modules/esm" context.
       esmSourceMapCache.set(filename, {
-        url: match.groups.sourceMappingURL,
-        data: dataFromUrl(basePath, match.groups.sourceMappingURL)
+        lineLengths: lineLengths(content),
+        data,
+        url
       });
     }
   }
@@ -71,6 +74,15 @@ function dataFromUrl(basePath, sourceMappingURL) {
     const sourceMapFile = resolve(basePath, sourceMappingURL);
     return sourceMapFromFile(sourceMapFile);
   }
+}
+
+// Cache the length of each line in the file that a source map was extracted
+// from. This allows translation from byte offset V8 coverage reports,
+// to line/column offset Source Map V3.
+function lineLengths(content) {
+  return content.split('\n').map((line) => {
+    return line.length;
+  });
 }
 
 function sourceMapFromFile(sourceMapFile) {
@@ -161,55 +173,13 @@ function appendCJSCache(obj) {
     const value = cjsSourceMapCache.get(Module._cache[key]);
     if (value) {
       obj[`file://${key}`] = {
-        url: value.url,
-        data: value.data
+        lineLengths: value.lineLengths,
+        data: value.data,
+        url: value.url
       };
     }
   });
 }
-
-// Create a prettified stacktrace, inserting context from source maps
-// if possible.
-const ErrorToString = Error.prototype.toString; // Capture original toString.
-const prepareStackTrace = (globalThis, error, trace) => {
-  // API for node internals to override error stack formatting
-  // without interfering with userland code.
-  // TODO(bcoe): add support for source-maps to repl.
-  if (overrideStackTrace.has(error)) {
-    const f = overrideStackTrace.get(error);
-    overrideStackTrace.delete(error);
-    return f(error, trace);
-  }
-
-  const { SourceMap } = require('internal/source_map/source_map');
-  const errorString = ErrorToString.call(error);
-
-  if (trace.length === 0) {
-    return errorString;
-  }
-  const preparedTrace = trace.map((t, i) => {
-    let str = i !== 0 ? '\n    at ' : '';
-    str = `${str}${t}`;
-    try {
-      const sourceMap = findSourceMap(t.getFileName(), error);
-      if (sourceMap && sourceMap.data) {
-        const sm = new SourceMap(sourceMap.data);
-        // Source Map V3 lines/columns use zero-based offsets whereas, in
-        // stack traces, they start at 1/1.
-        const [, , url, line, col] =
-                   sm.findEntry(t.getLineNumber() - 1, t.getColumnNumber() - 1);
-        if (url && line !== undefined && col !== undefined) {
-          str +=
-            `\n        -> ${url.replace('file://', '')}:${line + 1}:${col + 1}`;
-        }
-      }
-    } catch (err) {
-      debug(err.stack);
-    }
-    return str;
-  });
-  return `${errorString}\n    at ${preparedTrace.join('')}`;
-};
 
 // Attempt to lookup a source map, which is either attached to a file URI, or
 // keyed on an error instance.
@@ -230,8 +200,8 @@ function findSourceMap(uri, error) {
 }
 
 module.exports = {
+  findSourceMap,
   maybeCacheSourceMap,
-  prepareStackTrace,
   rekeySourceMap,
   sourceMapCacheToObject,
 };
